Move `Ignore` to dedicated module
